fix current sreach function
