--- test/unit/search_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class SearchTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end
